pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.9.0",
    "isort>=6.1.0",
    "langchain-google-genai>=2.0.10",
    "langchain-ollama>=0.3.10",
    "python-dotenv>=1.1.1",
]
